requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
confluent-kafka==2.2.0
Flask==2.3.3
Flask-Cors==4.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
nltk==3.8.1
numpy==1.25.2
pandas==2.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
regex==2023.8.8
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
textblob==0.17.1
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.4
vaderSentiment==3.3.2
Werkzeug==2.3.7
